requirements-lint.txt
# Linting and code quality tools with compatible versions
black==23.11.0
isort==5.12.0
flake8==5.0.4
flake8-django==1.1.5
flake8-docstrings==1.7.0
flake8-bugbear==23.3.12
mypy==1.7.1
django-stubs==4.2.7
djangorestframework-stubs==3.14.4
pylint==2.17.7
pylint-django==2.5.5
bandit==1.7.5
safety==2.3.5
pre-commit==3.5.0
packaging==21.3
pycodestyle==2.9.1
pyflakes==2.5.0
mccabe==0.7.0
autoflake==2.0.0